Separate out `core` and `api` logic. Make standalone "script" files for the web_app and cli.
